Move useful functions to general utils
